Implement generic streaming indexer for any enum
The indexer can be given any enumerable that contains leaderboard entry
keys sorted by score. It will take one key at a time and buffer them
in memory, while keeping track of some metrics (such as which chunk
we're currently in, and how many keys have we seen so far). As soon as a
score changes between current and previous key, it triggers the "flush"
of current buffer, and a start of a new, empty one. The flush itself
takes each key we buffered up, turns them into proper leaderboard
records, and returns a stream that emits each record. Finally, flushed
streams are concatenated so that they appear to be a single continuous 
stream on output.

You can describe the functionality like this.

1. You ask for the top leaderboard record.
2. Your request hits the concat() stream, and it tries to get the first
   item, so it tells `chunk_while()` to give it the first item.
3. The `chunk_while()` starts grabbing passed-in keys one at a time,
   while the score is the same among them.
4. As soon `chunk_while()` sees a first different (lower) score, it will
   immediately wrap the first chunk (all records of top rank) into a
   stream that formats each key into a valid leaderboard record, and
   yield this stream to concat().
5. Concat will take the first item of this stream and return to you.

At this point, if you ask for next records, concat will have them
readily available if they're still from the same rank. It will only go 
back and ask for the next chunk when you cross the rank boundary.