Refactor state handling and add TermStore
From now on storage behaviour functions are expected to return a state
that will be stored in the leaderboard struct. This was not needed for
EtsStore since the state is just a table/process name that always stays
the same. However, for TermStore, where the entire leaderboard is in a
variable, as a simple struct, we do need to return and overwrite state.

EtsStore therefore was rewritten to just always return name as its
state, while the new TermStore takes full advantage of this contract.